backend/requirements.txt
Pillow==10.3.0
numpy==1.22.0
pymongo==4.6.3
piexif>=1.1
werkzeug>=2.0
Flask>=2.0
Flask-PyMongo>=2.3
hashid